Move into header.
